backend/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-appconf==1.0.6
django-imagekit==5.0.0
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
packaging==24.0
pilkit==3.0
pillow==10.3.0
pytz==2024.1
PyYAML==6.0.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
